+++ CHANGELOG.md
@@ -42,7 +42,8 @@ stability. Since we are not ready for the 1.0 release yet, this does not matter 
 - Made `JoinGraph` and related objects available in the `optimizer` package
 
 ### 📰 Updates
-- Added a `simplify()` method to query predicates to provide all simplified counterparts of the joins and filters
+- Added a `simplify()` method to query predicates to provide all simplified counterparts of the joins and filters.
+  `all_simple()` can be used to check whether all predicates can be simplified beforehand.
 - Postgres database classes now have hashing and equality comparison support
 - Added _str_ support for all pipelines and stages
 - Improved type hints for `simplify()` and `enlist()`